% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_config_credentials.R
\name{config_set_creds}
\alias{config_set_creds}
\title{Set credentials for services in the package}
\usage{
config_set_creds(
  key_name = NULL,
  key_value = NULL,
  cache = FALSE,
  renv_path = ".Renviron"
)
}
\arguments{
\item{key_name}{The name of the environment variable to set (e.g., "WILDERLAB_KEY")}

\item{key_value}{The value to set for the environment variable}

\item{cache}{Logical indicating whether to save the key in the .Renviron file}

\item{renv_path}{The path to the .Renviron file (default is ".Renviron" in the working directory)}
}
\value{
None. Sets the environment variable and optionally updates .Renviron.
}
\description{
Sets a credential in the environment and optionally cache it in .Renviron
}
\examples{
# .Renviron will be created/updated in the working directory
\dontrun{
config_set_creds("WILDERLAB_KEY", "your_api_key_here", cache = TRUE)
}

}
