% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_config_credentials.R
\name{config_set_wilderlab_creds}
\alias{config_set_wilderlab_creds}
\title{Specifically set Wilderlab credentials}
\usage{
config_set_wilderlab_creds(
  wilderlab_key = NULL,
  wilderlab_secret = NULL,
  wilderlab_xapikey = NULL,
  cache = FALSE
)
}
\arguments{
\item{wilderlab_key}{The Wilderlab API key}

\item{wilderlab_secret}{The Wilderlab API secret}

\item{wilderlab_xapikey}{The Wilderlab X-API-Key}

\item{cache}{Logical indicating whether to save the keys in the .Renviron file}
}
\value{
None. Sets the environment variables and optionally updates .Renviron.
}
\description{
Sets Wilderlab API credentials in the environment and optionally cache them in .Renviron
}
\examples{
# .Renviron will be created/updated in the working directory
\dontrun{
config_set_wilderlab_creds(
  wilderlab_key = "your_key_here",
  wilderlab_secret = "your_secret_here",
  wilderlab_xapikey = "your_xapikey_here",
  cache = TRUE
)
}

}
\seealso{
\code{\link[=config_set_creds]{config_set_creds()}}
}
