% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_eml_base.R
\name{eml_add_contact}
\alias{eml_add_contact}
\title{Add a contact to the EML object}
\usage{
eml_add_contact(eml, given_name, family_name, email_address = NULL)
}
\arguments{
\item{eml}{EML object}

\item{given_name}{Given name of the contact}

\item{family_name}{Family name of the contact}

\item{email_address}{Email address of the contact (optional)}
}
\value{
EML object with added contact
}
\description{
Sets a contact person for the dataset. Multiple contacts can be added.
}
\examples{
eml <- eml_base() |>
eml_add_contact(
  given_name = "Jane",
  family_name = "Smith",
  email_address = "jsmith@email.com"
)

}
