% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_eml_base.R
\name{eml_add_creator}
\alias{eml_add_creator}
\title{Add a creator to the EML object}
\usage{
eml_add_creator(eml, given_name, family_name, email_address = NULL)
}
\arguments{
\item{eml}{EML object}

\item{given_name}{Given name of the creator}

\item{family_name}{Family name of the creator}

\item{email_address}{Email address of the creator (optional)}
}
\value{
EML object with added creator
}
\description{
Sets a creator person for the dataset. Multiple creators can be added.
}
\examples{
eml <- eml_base() |>
eml_add_creator(
  given_name = "Alice",
  family_name = "Johnson"
)

}
