% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_eml_base.R
\name{eml_add_persons}
\alias{eml_add_persons}
\title{Add multiple project persons to EML object}
\usage{
eml_add_persons(
  eml,
  type,
  persons = list(person()),
  parent_node = ".//dataset"
)
}
\arguments{
\item{eml}{EML object}

\item{type}{The type of people being added, "creator", "contact", or "personnel" for projects}

\item{persons}{A list of personnel, each created using the \code{person()} function}

\item{parent_node}{Which EML to insert under, should always be ".//dataset", unless adding project personnel, in which case it is ".//dataset/project"}
}
\value{
EML object with added project personnels
}
\description{
Adds multiple persons to an EML object.
}
\examples{
eml <- eml_base() |>
eml_set_project(
  title = "Test Project Title"
) |>
eml_add_persons(
  type = "personnel",
  persons = list(
    person(given = "Emily", family = "Davis"),
    person(given = "Michael", family = "Brown", email = "mb@email.com")
  )
)
print(eml)

}
