% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_eml_base.R
\name{eml_add_project_personnel}
\alias{eml_add_project_personnel}
\title{Add project personnel to EML object}
\usage{
eml_add_project_personnel(eml, given_name, family_name, email_address = NULL)
}
\arguments{
\item{eml}{EML object}

\item{given_name}{Given name of the personnel}

\item{family_name}{Family name of the personnel}

\item{email_address}{Email address of the personnel (optional)}
}
\value{
EML object with added project personnel
}
\description{
Adds personnel to the project section of the dataset. Multiple personnel can be added.
The project must be set before adding personnel.
}
\examples{
eml <- eml_base() |>
eml_set_project(
 title = "Test Project Title"
) |>
eml_add_project_personnel(
 given_name = "Emily",
 family_name = "Davis",
)
print(eml)

}
