% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_eml_base.R
\name{eml_base}
\alias{eml_base}
\title{Create a base EML object for a pipe}
\usage{
eml_base()
}
\value{
EML object
}
\description{
eml_* functions can be used in a pipe to build up an EML object, which is
an xml2 object representing the EML metadata for a Darwin Core Archive. This function
creates the base EML object with the required namespaces and dataset node.
}
\details{
When using these functions manually, please note that the order in which nodes are
added is important. Confer with existing EML files to ensure proper structure.
}
\examples{
eml <- eml_base()
print(eml)


}
