% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_eml_base.R
\name{eml_calculate_bounding_box}
\alias{eml_calculate_bounding_box}
\title{Calculate bounding box from a list of coordinates}
\usage{
eml_calculate_bounding_box(lats = NULL, longs = NULL)
}
\arguments{
\item{lats}{Numeric vector of latitudes}

\item{longs}{Numeric vector of longitudes}
}
\value{
A list with west, east, north, and south bounding coordinates
}
\description{
Returns the max and min lat/longs for use in setting a coverage object
}
\examples{
lats <- c(34.0, 35.0, 36.5)
longs <- c(-120.0, -119.5, -121.0)
bbox <- eml_calculate_bounding_box(lats, longs)
print(bbox)

}
