% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_eml_orchestrate.R
\name{eml_gbif_create}
\alias{eml_gbif_create}
\title{Create a minimally valid EML file for GBIF}
\usage{
eml_gbif_create(
  title,
  creators,
  abstract,
  license,
  coverage_geo_description,
  coverage_bbox,
  coverage_start_date,
  coverage_end_date,
  maintenance_frequency,
  maintenance_description,
  contacts,
  step_description,
  study_extent,
  sampling_description,
  project_title,
  project_personnel
)
}
\arguments{
\item{title}{The title of the dataset.}

\item{creators}{A list of person objects representing the dataset creators.}

\item{abstract}{A brief abstract describing the dataset.}

\item{license}{The license under which the dataset is released (in SPDX format).}

\item{coverage_geo_description}{A description of the geographic coverage of the dataset.}

\item{coverage_bbox}{A list or data frame containing the bounding box coordinates with
columns: west, east, north, south.}

\item{coverage_start_date}{The start date of data collection (YYYY-MM-DD).}

\item{coverage_end_date}{The end date of data collection (YYYY-MM-DD).}

\item{maintenance_frequency}{The frequency of metadata maintenance (e.g., "monthly").}

\item{maintenance_description}{A description of how the metadata is maintained.}

\item{contacts}{A list of person objects representing the dataset contacts.}

\item{step_description}{A description of the methods used in data collection.}

\item{study_extent}{A description of the study extent.}

\item{sampling_description}{A description of the sampling methods.}

\item{project_title}{The title of the project associated with the dataset.}

\item{project_personnel}{A list of person objects representing the project personnel.}
}
\value{
A xml2 object representing the EML metadata.
}
\description{
This function generates a basic EML (Ecological Metadata Language) file
required for publishing datasets to GBIF (Global Biodiversity Information Facility).
}
\examples{
creators <- list(
  person("Jane", "Smith", email = "jsmith@email.com")
)
contacts <- list(
  person("John", "Doe", email = "jd@email.com")
)
project_personnel <- list(
  person("Emily", "Davis"),
  person("Michael", "Brown")
)
bbox <- list(west = -120.0, east = -119.0, north = 35.0, south = 34.0)
eml <- eml_gbif_create(
  title = "Sample Dataset",
  creators = creators,
  abstract = "This dataset contains sample data for demonstration purposes.",
  license = "CC-BY-4.0",
  coverage_geo_description = "Sample Location",
  coverage_bbox = bbox,
  coverage_start_date = "2020-01-01",
  coverage_end_date = "2020-12-31",
  maintenance_frequency = "monthly",
  maintenance_description = "Metadata is updated monthly.",
  contacts = contacts,
  step_description = "Data collection methods description.",
  study_extent = "Study extent description.",
  sampling_description = "Sampling methods description.",
  project_title = "Sample Project",
  project_personnel = project_personnel
)

}
