% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_eml_orchestrate.R
\name{eml_gbif_template_create}
\alias{eml_gbif_template_create}
\title{Create a CSV template for GBIF EML metadata}
\usage{
eml_gbif_template_create(path)
}
\arguments{
\item{path}{The file path where the CSV template will be saved.}
}
\value{
None. The function writes a CSV file to the specified path.
}
\description{
This function generates a CSV file template that users can fill out
with their dataset metadata. The filled CSV can then be read and converted
into a valid EML file for GBIF using the \code{eml_gbif_template_read} function.
}
\examples{
\dontrun{
eml_gbif_template_create("gbif_eml_template.csv")
}


}
\seealso{
\code{\link[=eml_gbif_template_read]{eml_gbif_template_read()}}
}
