% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_eml_orchestrate.R
\name{eml_parse_person}
\alias{eml_parse_person}
\title{Parse a person string in git format into a structured list}
\usage{
eml_parse_person(input)
}
\arguments{
\item{input}{A string representing a person in git format.}
}
\value{
A list with elements: given_name, family_name, email, title.
}
\description{
This function takes a string representing a person in git format with an optional org/title,
Eg: \verb{John Smith <jsmith@email.com> (Organization, Senior Developer)} and parses it
into a structured list containing given name, family name, email, and title.
}
\examples{
input <- "John Smith <jsmith@email.com> (Senior Developer)"
person <- eml_parse_person(input)
print(person)

}
