% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_eml_base.R
\name{eml_set_abstract}
\alias{eml_set_abstract}
\title{Set abstract in EML object}
\usage{
eml_set_abstract(eml, abstract)
}
\arguments{
\item{eml}{EML object}

\item{abstract}{Abstract string}
}
\value{
EML object with updated abstract
}
\description{
Set the abstract of the dataset, i.e. for everything in a single Darwin Core Archive.
}
\examples{
eml <- eml_base() |>
eml_set_abstract("This dataset contains sample data.")
print(eml)

}
