% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_eml_base.R
\name{eml_set_license}
\alias{eml_set_license}
\title{Set license in EML object}
\usage{
eml_set_license(eml, license_identifier = "CCBYNC4.0")
}
\arguments{
\item{eml}{EML object}

\item{license_identifier}{License identifier string (optional)}

\item{license_name}{License name string}

\item{license_url}{License URL string (optional)}
}
\value{
EML object with updated license
}
\description{
Set the license of the dataset, choosing from supported licenses.
}
\examples{
eml <- eml_base() |>
eml_set_license("CC-BY-NC-4.0")
print(eml)

}
