% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_eml_base.R
\name{eml_set_maintenance}
\alias{eml_set_maintenance}
\title{Set maintenance in EML object}
\usage{
eml_set_maintenance(eml, description = "Not specified.", frequency = "Unknown")
}
\arguments{
\item{eml}{EML object}

\item{description}{Maintenance description string}

\item{frequency}{Maintenance update frequency string}
}
\value{
EML object with updated maintenance
}
\description{
Sets the maintenance information of the dataset.
}
\examples{
eml <- eml_base() |>
eml_set_maintenance(
 description = "This metadata is maintained regularly.",
 frequency = "monthly"
)
print(eml)

}
