% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_eml_base.R
\name{eml_set_methods}
\alias{eml_set_methods}
\title{Set methods in EML object}
\usage{
eml_set_methods(
  eml,
  step_description = "No methods provided.",
  study_extent = "Not specified.",
  sampling_description = "Not specified."
)
}
\arguments{
\item{eml}{EML object}

\item{step_description}{Step description string}

\item{study_extent}{Study extent description string}

\item{sampling_description}{Sampling description string}
}
\value{
EML object with updated methods
}
\description{
Sets the methods information of the dataset.
}
\examples{
eml <- eml_base() |>
eml_set_methods(
  step_description = "This is a test methods description.",
  study_extent = "Study extent description.",
  sampling_description = "Sampling description."
)
print(eml)

}
