% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_eml_base.R
\name{eml_set_project}
\alias{eml_set_project}
\title{Set project in EML object}
\usage{
eml_set_project(eml, title)
}
\arguments{
\item{eml}{EML object}

\item{title}{Project title string}
}
\value{
EML object with updated project
}
\description{
Sets the project information of the dataset.
}
\examples{
eml <- eml_base() |>
eml_set_project(
 title = "Sample Project Title"
)
print(eml)

}
