% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_eml_base.R
\name{eml_set_title}
\alias{eml_set_title}
\title{Set title in EML object}
\usage{
eml_set_title(eml, title)
}
\arguments{
\item{eml}{EML object}

\item{title}{Title string}
}
\value{
EML object with updated title
}
\description{
Set the title of the dataset, i.e. for everything in a single Darwin Core Archive.
}
\examples{
eml <- eml_base() |>
 eml_set_title("My Sample Dataset")
print(eml)

}
