% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_zip.R
\name{gen_make_dwc_archive}
\alias{gen_make_dwc_archive}
\title{Generate Darwin Core Archive ZIP}
\usage{
gen_make_dwc_archive(dwc_tibbles, eml_xml, meta_xml, path)
}
\arguments{
\item{dwc_tibbles}{A named list of tibbles representing Darwin Core data tables.}

\item{eml_xml}{An XML object representing the EML metadata.}

\item{meta_xml}{An XML object representing the Darwin Core Archive metadata.}

\item{path}{The file path where the resulting ZIP archive will be saved.}
}
\value{
The file path of the created Darwin Core Archive ZIP file.
}
\description{
Generates a Darwin Core Archive ZIP file containing the provided data tables and metadata XML files.
}
\examples{
# Requires an internet connection to fetch latest DwC schema
\dontrun{
# Example Darwin Core tibbles
dwc_tibbles <- list(
  occurrence_core = tibble::tibble(
    id = c("1", "2"),
    scientificName = c("Puma concolor", "Canis lupus"),
    eventDate = c("2020-01-01", "2020-11-20")
  ),
  dna_extension = tibble::tibble(
    id = c("1", "2"),
    DNA_sequence = c("ATCG", "GCTA")
  ),
  mof_extension = tibble::tibble(
    id = c("1", "2"),
    measurementType = c("Length", "Weight"),
    measurementValue = c(15.5, 22.3)
  )
)
# Example EML and metadata XML
eml_xml <- xml2::read_xml("<eml></eml>")
meta_xml <- xml2::read_xml("<archive></archive>")
# Generate Darwin Core Archive ZIP
dwc_zip_path <- gen_make_dwc_archive(
  dwc_tibbles = dwc_tibbles,
  eml_xml = eml_xml,
  meta_xml = meta_xml,
  path = "dwc_archive.zip"
)
print(dwc_zip_path)
}


}
