% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_dataframes.R
\name{gen_tibble_to_dwc_frames}
\alias{gen_tibble_to_dwc_frames}
\title{Split dataframe by Darwin Core file}
\usage{
gen_tibble_to_dwc_frames(data)
}
\arguments{
\item{data}{A tibble containing the data to be converted.}
}
\value{
A named list of tibbles, each representing a Darwin Core extension.
}
\description{
Converts a singular tibble into a list of Darwin Core data frames.
Each data frame in the list corresponds to a Darwin Core extension.
}
\examples{
# Requires an internet connection to fetch latest DwC schema
\dontrun{
# Example tibble
df <- tibble::tibble(
  scientificName = c("Puma concolor", "Canis lupus"),
  eventDate = c("2020-01-01", "2020-11-20"),
  decimalLatitude = c(34.05, 36.16),
  DNA_sequence = c("ATCG", "GCTA"),
  measurementType = c("Length", "Weight"),
  measurementValue = c(15.5, 22.3)
)
# Convert to Darwin Core frames
dwc_frames <- gen_tibble_to_dwc_frames(df)
print(names(dwc_frames))
# [1] "occurrence_core" "dna_extension"   "mof_extension"
}


}
