% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_config_dwc_schema.R
\name{get_all_gbif_dwc_schema}
\alias{get_all_gbif_dwc_schema}
\title{Get latest GBIF Darwin Core schema}
\usage{
get_all_gbif_dwc_schema(bind = FALSE)
}
\arguments{
\item{bind}{Logical indicating whether to bind all schemas into a single data frame (default: FALSE)}
}
\value{
A list of data frames for each schema or a single combined data frame if bind is TRUE
}
\description{
Fetches the latest versions of the occurrence core, DNA extension, and measurements or facts extension schemas
Uses the GBIF schema repository URLs to get the most recent files
}
\examples{
# Get all schemas as a list of data frames
# Requires an internet connection
\dontrun{
schemas_list <- get_all_gbif_dwc_schema()
print(names(schemas_list))
# [1] "occurrence_core" "dna_extension"   "mof_extension" 
}

}
