% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_upload_resource.R
\name{ipt_publish_resource}
\alias{ipt_publish_resource}
\title{Publish IPT resource}
\usage{
ipt_publish_resource(session = NULL, shortname)
}
\arguments{
\item{session}{An optional IPT session object. If NULL, a new session will be created.}

\item{shortname}{The shortname of the resource to be published.}
}
\value{
An updated IPT session object after publishing the resource.
}
\description{
Publishes a resource on the Integrated Publishing Toolkit (IPT) using its shortname.
Note that this triggers the publication process, but does not verify if the resource
is successfully published, and the IPT will need to process the request before it
appears as published.
}
\examples{
\dontrun{
shortname <- "my_resource"
ipt_login() |>
ipt_publish_resource(
 shortname = shortname
)
# Check IPT after a few seconds to confirm publication
}

}
\seealso{
\code{\link[=ipt_login]{ipt_login()}}, \code{\link[=ipt_get_jsessionid]{ipt_get_jsessionid()}}, \code{\link[=ipt_upload_archive]{ipt_upload_archive()}}, \code{\link[=ipt_save_metadata]{ipt_save_metadata()}}
}
