% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_upload_resource.R
\name{ipt_register_resource}
\alias{ipt_register_resource}
\title{Register a resource with GBIF}
\usage{
ipt_register_resource(session = NULL, shortname)
}
\arguments{
\item{session}{An optional IPT session object. If NULL, a new session will be
created.}

\item{shortname}{The shortname of the resource to be registered with GBIF.}
}
\value{
An updated IPT session object after registering the resource.
}
\description{
Registers a resource on GBIF using its shortname in the IPT.
}
\examples{
\dontrun{
shortname <- "my_resource"
ipt_login() |>
ipt_register_with_gbif(
  shortname = shortname
)
}

}
\seealso{
\code{\link[=ipt_login]{ipt_login()}}, \code{\link[=ipt_get_jsessionid]{ipt_get_jsessionid()}}
}
