% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_upload_resource.R
\name{ipt_save_metadata}
\alias{ipt_save_metadata}
\title{Trigger metadata save on IPT resource}
\usage{
ipt_save_metadata(session = NULL, shortname, org = NULL)
}
\arguments{
\item{session}{An optional IPT session object. If NULL, a new session will be created.}

\item{shortname}{The shortname of the resource whose metadata is to be saved.}

\item{org}{An optional organization name to set for the resource.}
}
\value{
An updated IPT session object after saving the metadata.
}
\description{
Even if an EML file in a Darwin Core Archive contains all required metadata,
the IPT does not automatically recognize the data as complete. This function
navigates to the metadata editing page for the specified resource and triggers
a save action to ensure the IPT acknowledges the metadata. It sets the shortname
as part of the form submission, to trigger the save.
}
\examples{
\dontrun{
shortname <- "my_resource"
ipt_login() |>
ipt_save_metadata(
  shortname = shortname
)
}

}
\seealso{
\code{\link[=ipt_login]{ipt_login()}}, \code{\link[=ipt_get_jsessionid]{ipt_get_jsessionid()}}, \code{\link[=ipt_upload_archive]{ipt_upload_archive()}}, \code{\link[=ipt_publish_resource]{ipt_publish_resource()}}
}
