% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_upload_resource.R
\name{ipt_set_visible}
\alias{ipt_set_visible}
\title{Make IPT resource publicly visible}
\usage{
ipt_set_visible(session = NULL, shortname)
}
\arguments{
\item{session}{An optional IPT session object. If NULL, a new session will be created.}

\item{shortname}{The shortname of the resource to be made public.}
}
\value{
An updated IPT session object after making the resource public.
}
\description{
Sets the specified IPT resource to be publicly visible.
}
\examples{
\dontrun{
shortname <- "my_resource"
ipt_login() |>
ipt_set_visible(
  shortname = shortname
)
# Check IPT after a few seconds to confirm visibility
}

}
\seealso{
\code{\link[=ipt_login]{ipt_login()}}, \code{\link[=ipt_get_jsessionid]{ipt_get_jsessionid()}}
}
