% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_upload_resource.R
\name{ipt_upload_and_publish}
\alias{ipt_upload_and_publish}
\title{Upload and publish Darwin Core Archive to IPT}
\usage{
ipt_upload_and_publish(archive, shortname)
}
\arguments{
\item{archive}{The file path to the Darwin Core Archive ZIP file.}

\item{shortname}{The shortname for the new resource in the IPT.}
}
\value{
TRUE if all steps are completed successfully.
}
\description{
This function combines the steps of logging into the IPT, uploading a Darwin Core Archive,
saving the metadata, and publishing the resource. It takes the path to the archive
and the desired shortname as inputs, and performs all necessary actions in sequence.
}
\examples{
\dontrun{
archive_path <- "path/to/dwca.zip"
shortname <- "my_resource"
ipt_upload_and_publish(
 archive = archive_path,
 shortname = shortname
)
# Check IPT after a few seconds to confirm publication
}

}
\seealso{
\code{\link[=ipt_login]{ipt_login()}}, \code{\link[=ipt_upload_archive]{ipt_upload_archive()}}, \code{\link[=ipt_save_metadata]{ipt_save_metadata()}}, \code{\link[=ipt_publish_resource]{ipt_publish_resource()}}
}
