% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_upload_resource.R
\name{ipt_upload_archive}
\alias{ipt_upload_archive}
\title{Upload Darwin Core Archive to IPT}
\usage{
ipt_upload_archive(session = NULL, archive, shortname)
}
\arguments{
\item{session}{An optional IPT session object. If NULL, a new session will be created.}

\item{archive}{The file path to the Darwin Core Archive ZIP file.}

\item{shortname}{The shortname for the new resource in the IPT.}
}
\value{
An updated IPT session object after the upload.
}
\description{
Uploads a Darwin Core Archive (DwCA) to the Integrated Publishing Toolkit (IPT)
and creates a new resource with the specified shortname.
}
\examples{
\dontrun{
archive_path <- "path/to/dwca.zip"
shortname <- "my_resource"
ipt_login() |>
ipt_upload_archive(
 archive = archive_path,
 shortname = shortname
)
}

}
\seealso{
\code{\link[=ipt_login]{ipt_login()}}, \code{\link[=ipt_get_jsessionid]{ipt_get_jsessionid()}}, \code{\link[=ipt_save_metadata]{ipt_save_metadata()}}, \code{\link[=ipt_publish_resource]{ipt_publish_resource()}}
}
