% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_meta.R
\name{meta_write_xml}
\alias{meta_write_xml}
\title{Write Darwin Core Archive metadata XML to file}
\usage{
meta_write_xml(meta_xml, path)
}
\arguments{
\item{meta_xml}{Darwin Core Archive metadata XML object}

\item{path}{File path to write the metadata XML}
}
\value{
The file path where the metadata XML was written
}
\description{
Wrapper function to write the Darwin Core Archive metadata XML to a specified file path.
}
\examples{
# Requires an internet connection to fetch latest DwC schema
\dontrun{
# Example data frame
df <- data.frame(
  id = c("uuid1", "uuid2"),
  scientificName = c("Puma concolor", "Canis lupus"),
  eventDate = c("2020-01-01", "2020-11-20")
)
# Create metadata XML with core file
meta_xml <- meta_xml_base() |>
meta_xml_add_file(
  type="core",
  df=df,
  location="occurrence_core.txt"
)
# Write metadata XML to file
meta_path <- meta_write_xml(meta_xml, "meta.xml")
print(paste("Metadata XML written to:", meta_path))
}

}
