% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_meta.R
\name{meta_xml_add_file}
\alias{meta_xml_add_file}
\title{Add a file to Darwin Core Archive metadata XML}
\usage{
meta_xml_add_file(
  xml,
  type,
  df,
  location,
  encoding = "UTF-8",
  fieldsTerminatedBy = "\\\\t",
  linesTerminatedBy = "\\\\n",
  fieldsEnclosedBy = "",
  ignoreHeaderLines = "1",
  rowType = "http://rs.tdwg.org/dwc/terms/Occurrence"
)
}
\arguments{
\item{xml}{Darwin Core Archive metadata XML object}

\item{type}{Type of file to add (e.g., "core" or "extension")}

\item{df}{Data frame to describe}

\item{location}{Name of the data file}

\item{encoding}{File encoding (default: "UTF-8")}

\item{fieldsTerminatedBy}{Character that separates fields (default: tab)}

\item{linesTerminatedBy}{Character that separates lines (default: newline)}

\item{fieldsEnclosedBy}{Character that encloses fields (default: double quote)}

\item{ignoreHeaderLines}{Number of header lines to ignore (default: 1)}

\item{rowType}{Row type URI (default: occurrence)}
}
\value{
Updated Darwin Core Archive metadata XML object
}
\description{
The function adds a core or extension file description to the Darwin Core Archive metadata XML.
It includes attributes for file encoding, field delimiters, and row types.
It also maps the columns of the provided data frame to Darwin Core terms based on the GBIF schema.
Downstream functions will make tab delimited text files, so the defaults should only be changed
if you are designing a custom workflow.
}
\examples{
# Requires an internet connection to fetch latest DwC schema
\dontrun{
# Example data frame
df <- data.frame(
 id = c("uuid1", "uuid2"),
 scientificName = c("Puma concolor", "Canis lupus"),
 eventDate = c("2020-01-01", "2020-11-20")
)
# Add core file to metadata XML
meta_xml <- meta_xml_base() |>
meta_xml_add_file(
 type="core",
 df=df,
 location="occurrence_core.txt"
)
print(xml2::as_xml_document(meta_xml))
}

}
