% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_generate_dwc_meta.R
\name{meta_xml_base}
\alias{meta_xml_base}
\title{Generate base Darwin Core Archive metadata XML}
\usage{
meta_xml_base()
}
\value{
Darwin Core Archive metadata XML object
}
\description{
Automatically sets up an empty core node for adding fields later
}
\examples{
meta_xml <- meta_xml_base()
print(xml2::as_xml_document(meta_xml))
}
