% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_validate_field_scientificName.R
\name{val_check_name_list}
\alias{val_check_name_list}
\title{Check multiple scientific names against the GBIF Backbone Taxonomy}
\usage{
val_check_name_list(name_list, quiet = FALSE)
}
\arguments{
\item{name_list}{A vector of scientific names to check}
}
\value{
A data frame with scientific names and a logical column indicating if a match was found
}
\description{
Iterates over a list of scientific names and checks each one against the GBIF Backbone Taxonomy
using the GBIF Species API. Returns a data frame indicating which names were found.
The underlying API requests are performed in parallel for efficiency, and will show a progress bar
unless \code{quiet = TRUE}.
}
\details{
Requests will be dynamically retried on failure, and throttled to 10 requests per second.
GBIF does not require an API key and has not specified strict rate limits, but this is to be respectful of their servers.
}
\examples{
\dontrun{
names_to_check <- c("Puma concolor", "Invalid species name")
result_df <- val_check_name_list(names_to_check)
print(result_df)
# A tibble: 2 × 2
#  scientificName          match
#  <chr>                   <lgl>
#1 Puma concolor          TRUE
#2 Invalid species name   FALSE
}


}
\seealso{
\code{\link[=val_check_scientific_name]{val_check_scientific_name()}}
}
