% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_validate_field_scientificName.R
\name{val_check_scientific_name}
\alias{val_check_scientific_name}
\title{Check if a scientific name exists in the GBIF Backbone Taxonomy}
\usage{
val_check_scientific_name(name)
}
\arguments{
\item{name}{A scientific name to check}
}
\value{
The number of matches found in the GBIF Backbone Taxonomy
}
\description{
This function queries the GBIF Species API to determine if the provided scientific name
exists in the GBIF Backbone Taxonomy.
}
\examples{
\dontrun{
val_check_scientific_name("Puma concolor") # should return TRUE
val_check_scientific_name("Invalid species name") # should return FALSE
}


}
\seealso{
\code{\link[=val_check_name_list]{val_check_name_list()}}
}
