% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_validate_orchestrator.R
\name{val_validate_all_fields}
\alias{val_validate_all_fields}
\title{Validate All Fields in Data Frame}
\usage{
val_validate_all_fields(df)
}
\arguments{
\item{df}{A data frame containing the data to be validated.}
}
\value{
A list of validation results, including any issues found during the validation process.
}
\description{
This function serves as an orchestrator to validate all fields in a given data frame.
It builds each validator class based on the column names of the data frame and applies the respective
validation methods to ensure data integrity.
}
\examples{
# Example data frame
df <- data.frame(
  scientificName = c("Puma concolor", "Invalid species name", "Canis lupus"),
  eventDate = c("2020-01-01", "2020-11-20", "Invalid Date")
)
# Validate all fields
issues <- val_validate_all_fields(df)
if (!is.null(issues)) {
  print(issues)
} else {
  print("No issues found")
}


}
