% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_validate_dwc_columns.R
\name{val_validate_dwc_to_level}
\alias{val_validate_dwc_to_level}
\title{Validates existence of DwC columns to a specified standard}
\usage{
val_validate_dwc_to_level(df, level)
}
\arguments{
\item{df}{A data frame to validate}

\item{level}{A character string specifying the Darwin Core level to validate against.
Must be one of "required", "recommended", or "preferred".}
}
\value{
A list of issue tibbles if any issues are found, or NULL if no issues are found
}
\description{
Given a data frame, checks for the presence of required, recommended, or preferred terms
as defined by the Darwin Core standard / GBIF data quality requirements.
}
\examples{
# Example data frame missing some recommended terms
# Requires internet connection to fetch latest DwC terms
\dontrun{
df <- data.frame(
 occurrenceID = c("occ1", "occ2"),
eventDate = c("2020-01-01", "2020-01-02"),
scientificName = c("Species A", "Species B"),
decimalLatitude = c(34.05, 36.16)
)
# Validate against recommended level
issues <- val_validate_dwc_to_level(df, level = "recommended")
if (!is.null(issues)) {
  print(issues)
} else {
  print("No issues found")
}
# Validate against preferred level
issues_preferred <- val_validate_dwc_to_level(df, level = "preferred")
if (!is.null(issues_preferred)) {
  print(issues_preferred)
} else {
  print("No issues found")
}
}


}
