% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ingest_wilderlab.R
\name{wl_export_wilderlab_data}
\alias{wl_export_wilderlab_data}
\title{Export Wilderlab data to local folder}
\usage{
wl_export_wilderlab_data(wilderlab_data, path)
}
\arguments{
\item{wilderlab_data}{A list containing data frames for jobs, samples, taxa, and records}

\item{path}{The path to the folder where the CSV files will be saved}
}
\value{
None. The function writes CSV files to the specified folder.
}
\description{
Creates jobs.csv, samples.csv, taxa.csv, and records.csv files in the specified folder.
Each file is formatted as it exists in the Wilderlab API.
}
\examples{
\dontrun{
# Assuming wilderlab_data is a list with jobs, samples, taxa, and records data frames
# Which can be obtained using wl_get_all_wilderlab_data()
wilderlab_data <- wl_get_all_wilderlab_data()
wl_export_wilderlab_data(wilderlab_data, path = "path/to/save/folder")
}

}
\seealso{
\code{\link[=wl_get_wilderlab_data]{wl_get_wilderlab_data()}}, \code{\link[=wl_get_all_wilderlab_data]{wl_get_all_wilderlab_data()}}
}
