% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ingest_wilderlab.R
\name{wl_get_all_wilderlab_data}
\alias{wl_get_all_wilderlab_data}
\title{Get all data from Wilderlab API}
\usage{
wl_get_all_wilderlab_data(public_only = TRUE)
}
\arguments{
\item{public_only}{Logical; if TRUE, only public data is retrieved from Wilderlab}
}
\value{
A list containing data frames for jobs, samples, taxa, and records
}
\description{
This function retrieves all data from the Wilderlab API, including jobs, samples, taxa, and all records in the job.
It handles authentication and returns the data as a list of data frames.
\emph{Use with caution!} Depending on the number of jobs and records, this may take a long time.
}
\examples{
\dontrun{
# Get all data from Wilderlab
all_data <- wl_get_all_wilderlab_data()
print(head(all_data$records))
}
}
