% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ingest_wilderlab.R
\name{wl_get_records_from_joblist}
\alias{wl_get_records_from_joblist}
\title{Get records data for multiple JobIDs from Wilderlab API}
\usage{
wl_get_records_from_joblist(jobIDs)
}
\arguments{
\item{jobIDs}{A vector of JobIDs for which to retrieve records data}
}
\value{
A data frame containing the combined records data for all specified JobIDs
}
\description{
This function retrieves records data for a list of JobIDs from the Wilderlab API.
It performs parallel requests for efficiency and combines the results into a single data frame.
}
\examples{
\dontrun{
# Get records data for multiple JobIDs
samples <- wl_get_wilderlab_data("samples")
jobIDs <- samples |> 
  filter(MakeDataPublic == 1) |> 
  pull(JobID) |> 
  unique()
records_data <- wl_get_records_from_joblist(jobIDs)
print(head(records_data))
}

}
