% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ingest_wilderlab.R
\name{wl_get_wilderlab_data}
\alias{wl_get_wilderlab_data}
\title{Get data from Wilderlab API}
\usage{
wl_get_wilderlab_data(table, jobID = NULL)
}
\arguments{
\item{table}{The name of the table to query (jobs, samples, taxa, records)}

\item{jobID}{Optional JobID to filter results (only for records tables)}

\item{request_only}{Logical; if TRUE, only the request object is returned without performing the request}
}
\value{
A data frame containing the requested data
}
\description{
This function retrieves data from the Wilderlab API for a specified table and optional JobID.
It displays terminal output and formats the response.
Underlying functions will request Wilderlab API keys if necessary. These are not stored
Unless you choose to cache them in a .Renviron, which will be made in the working directory
}
\examples{
# Get jobs data
\dontrun{
jobs_data <- wl_get_wilderlab_data("jobs")
print(head(jobs_data))
}

}
