% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ingest_wilderlab.R
\name{wl_get_wilderlab_taxa}
\alias{wl_get_wilderlab_taxa}
\title{Get Wilderlab taxa data}
\usage{
wl_get_wilderlab_taxa()
}
\value{
A data frame containing the taxa data
}
\description{
Taxa data have been migrated to a fixed location so this should be fetched directly,
instead of using the generic function. Note that the file is now in RDS format.
}
\examples{
\dontrun{
taxa_data <- wl_get_wilderlab_taxa()
print(head(taxa_data))
}

}
