% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ingest_wilderlab.R
\name{wl_map_wilderlab_data}
\alias{wl_map_wilderlab_data}
\title{Map Wilderlab data to standard column names}
\usage{
wl_map_wilderlab_data(wilderlab_data, mappings = wl_wilderlab_dwc_mappings())
}
\arguments{
\item{wilderlab_data}{A list containing data frames for jobs, samples, taxa, and records}

\item{mappings}{A named character vector where names are standard column names and values are Wilderlab column names}
}
\value{
A data frame with standardised column names
}
\description{
This function maps the columns of Wilderlab data frames to standard column names used in the darwin core
Also joins the data frames together into a single data frame using the records as the main table
Optionally, mappings can be left blank just to get the joined dataframe, purely for testing purposes
}
\examples{
\dontrun{
# Assuming wilderlab_data is a list with jobs, samples, taxa, and records data frames
# Which can be obtained using wl_get_all_wilderlab_data()
wilderlab_data <- wl_get_all_wilderlab_data()
mapped_data <- wl_map_wilderlab_data(wilderlab_data)
print(head(mapped_data))
}
}
