% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ingest_wilderlab.R
\name{wl_read_wilderlab_data}
\alias{wl_read_wilderlab_data}
\title{Read Wilderlab data from local folder}
\usage{
wl_read_wilderlab_data(folder_path)
}
\arguments{
\item{folder_path}{The path to the folder containing the CSV files}
}
\value{
A list containing data frames for jobs, samples, taxa, and records
}
\description{
This function reads Wilderlab data from CSV files in a specified folder.
It expects the folder to contain jobs.csv, samples.csv, taxa.csv, and records.csv
}
\examples{
\dontrun{
# Read Wilderlab data from a local folder
wilderlab_data <- wl_read_wilderlab_data("path/to/folder")
print(head(wilderlab_data$jobs))
}

}
\seealso{
\code{\link[=wl_export_wilderlab_data]{wl_export_wilderlab_data()}}
}
