% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_cli_interface.R
\name{eDNABridge}
\alias{eDNABridge}
\title{eDNABridge CLI Interface}
\usage{
eDNABridge()
}
\value{
Invisibly returns NULL after exiting the CLI interface.
}
\description{
This function launches a command-line interface (CLI) for the eDNABridge package,
allowing users to interactively configure and execute a workflow for publishing
eDNA data to GBIF. Users can select options for data ingestion, validation,
Darwin Core Archive generation, and data upload to the GBIF IPT.
}
\examples{
\dontrun{
eDNABridge()
}
}
